Added a layer of abstraction
The Transformer and LSTM classes are now renamed to TransformerEncoder and TwoLayerLSTM and contained in the new models.py. Other programs are changed to reflect this.
