+++ kitchen_pose_prediction/KitchenLSTMTest.py
@@ -7,10 +7,10 @@ import gc # This is an ugly hack
 import numpy as np
 import matplotlib.pyplot as plt
 import torch
-import torch.nn as nn
 from torch.autograd import Variable
 from sklearn.preprocessing import MinMaxScaler
 from ParseKitchenC3D import load_and_preprocess_mocap
+from Models import TwoLayerLSTM
 
 seq_length = 100
 predict_length = 1
@@ -46,33 +46,6 @@ def load_data(filename):
     
     return train_data, test_data
 
-class LSTM(nn.Module):
-    def __init__(self, hidden_layers=64, frame_dimension = 1):
-        super(LSTM, self).__init__()
-        self.hidden_layers = hidden_layers
-        self.input_size = frame_dimension
-        # lstm1, lstm2, linear are all layers in the network
-        self.lstm1 = nn.LSTMCell(frame_dimension, self.hidden_layers)
-        self.lstm2 = nn.LSTMCell(self.hidden_layers, self.hidden_layers)
-        self.linear = nn.Linear(self.hidden_layers, frame_dimension)
-        
-    def forward(self, y, pre_output_len=1):
-        outputs = []
-        h_t = torch.zeros(1, self.hidden_layers, dtype=torch.float32)
-        c_t = torch.zeros(1, self.hidden_layers, dtype=torch.float32)
-        h_t2 = torch.zeros(1, self.hidden_layers, dtype=torch.float32)
-        c_t2 = torch.zeros(1, self.hidden_layers, dtype=torch.float32)
-        
-        for i, frame in enumerate(y.split(1, dim=0)):
-            h_t, c_t = self.lstm1(frame, (h_t, c_t)) # initial hidden and cell states
-            h_t2, c_t2 = self.lstm2(h_t, (h_t2, c_t2)) # new hidden and cell states
-            output = self.linear(h_t2) # output from the last FC layer
-            if i >= pre_output_len: outputs.append(output)
-
-        # transform list to tensor    
-        outputs = torch.stack(outputs)
-        return outputs
-
 if __name__ == "__main__":
     num_epochs = 100
     learning_rate = 0.01
@@ -84,7 +57,7 @@ if __name__ == "__main__":
     
     num_classes = 198
     
-    lstm = LSTM(hidden_size, input_size)
+    lstm = TwoLayerLSTM(hidden_size, input_size)
     
     criterion = torch.nn.MSELoss()    # mean-squared error for regression
     optimizer = torch.optim.Adam(lstm.parameters(), lr=learning_rate)
